Issue 144: Added strafe and spin turning (optional rotate in place instead of moving forward) to handler.
I believe this makes the ThirdPersonHandler/ChaseCamera system complete.  I will make the latest features available in the Editor for testing and if all looks good, this issue can be marked complete.

Known functionality: strafe is defined as being at a right angle to the target direction or camera direction (depending on setting) so when using with camera relative movement, strafe commands will have the target instantly face camera forward which may look odd to some people.  It's not a bug though, it's known functionality that can be revisited if it's annoying.


git-svn-id: http://jmonkeyengine.googlecode.com/svn/trunk@2523 75d07b2b-3a1a-0410-a2c5-0572b91ccdca