Issue: MINOR performance update
Added shortcircuit in batch for CULL_NEVER to prevent needless camera checking similar to what I added to Spatial.  Also added shortcut on TriMesh for drawing batches...  if only 1 batch, any bounding intersection test on batch will be redundant to that done for trimesh, so skip to draw.  This returns us back much of that small fps loss incurred with introduction of batches.  :)


git-svn-id: http://jmonkeyengine.googlecode.com/svn/trunk@3020 75d07b2b-3a1a-0410-a2c5-0572b91ccdca