Issue minor:
Optimized getType() in TriangleBatch and SharedMesh. getType() is invoked a lot by the opaque renderqueue sorting, and used to call super.getType() for determining the type.

Of course, the result of this caluclation is entirely predictable (the class hierachy can't change run time). Replacing the super call allows the calculation to be done compile time.

One would hope that a smart VM would figure out that a method that returns a final value, can then be used to eliminate any further calculations in a method that uses the result only in calculations with other final values. The problem is, of course, you can't profile if this is happening. Some artificial tests seemed to suggest it's a bit faster now, so probably my VM just inlines the method or something like that..

Since in any case it won't be slower, and it'll mess up profiling results less,  I decided.. why not!
What? You're still reading this? wow ;)


git-svn-id: http://jmonkeyengine.googlecode.com/svn/trunk@3060 75d07b2b-3a1a-0410-a2c5-0572b91ccdca