Fixed a (possibly serious) performance issue with setting cursors.
The cursor cache used URL objects as keys, these access the network
every time they have equals called on them and, as a result, are
unsuitable for use as map keys. I have replaced the map keys with
strings based on the normalised URIs where possible, and the raw
externalised URLs where not. The public API is unchanged.

Forum Topic 9862.0

git-svn-id: http://jmonkeyengine.googlecode.com/svn/trunk@4055 75d07b2b-3a1a-0410-a2c5-0572b91ccdca